api/requirements.txt
# 星河智安 (XingHe ZhiAn) - AI安全攻击可视化平台
# Backend Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Redis and Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# AI/ML - PyTorch ecosystem
torch==2.5.1
torchvision==0.20.1
numpy==1.24.4
pillow==10.1.0
opencv-python==4.8.1.78

# YOLOv8 (Ultralytics)
ultralytics==8.0.206

# Image processing and visualization
matplotlib==3.7.4
seaborn==0.12.2
plotly==5.17.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional ML utilities
scikit-learn==1.3.2
tqdm==4.66.1
